# [Propagation] Improve the perf by passing structs by reference