… src/OpenTelemetry.Api/.publicApi/net462/PublicAPI.Unshipped.txt
… src/OpenTelemetry.Api/.publicApi/netstandard2.0/PublicAPI.Unshipped.txt
… src/OpenTelemetry.Extensions.Propagators/.publicApi/net462/PublicAPI.Unshipped.txt
… src/OpenTelemetry.Extensions.Propagators/.publicApi/netstandard2.0/PublicAPI.Unshipped.txt
… src/OpenTelemetry.Extensions.Propagators/B3Propagator.cs
… src/OpenTelemetry.Instrumentation.Http/Implementation/HttpHandlerDiagnosticListener.cs
… src/OpenTelemetry.Instrumentation.Http/Implementation/HttpWebRequestActivitySource.netfx.cs
… src/OpenTelemetry.Shims.OpenTracing/TracerShim.cs
… test/Benchmarks/Context/PropagationBenchmarks.cs
… test/Benchmarks/Instrumentation/AspNetCoreExtractionBenchmark.cs
… test/OpenTelemetry.Instrumentation.AspNet.Tests/HttpInListenerTests.cs
… test/OpenTelemetry.Instrumentation.AspNet.Tests/OpenTelemetry.Instrumentation.AspNet.Tests.csproj
… test/OpenTelemetry.Instrumentation.AspNetCore.Tests/BasicTests.cs
… test/OpenTelemetry.Instrumentation.AspNetCore.Tests/OpenTelemetry.Instrumentation.AspNetCore.Tests.csproj
… test/OpenTelemetry.Instrumentation.Grpc.Tests/GrpcTests.client.cs
… test/OpenTelemetry.Instrumentation.Grpc.Tests/OpenTelemetry.Instrumentation.Grpc.Tests.csproj
… test/OpenTelemetry.Instrumentation.Http.Tests/HttpClientTests.Basic.netcore31.cs
… test/OpenTelemetry.Instrumentation.Http.Tests/HttpWebRequestTests.Basic.netfx.cs
… test/OpenTelemetry.Instrumentation.Http.Tests/OpenTelemetry.Instrumentation.Http.Tests.csproj
… test/OpenTelemetry.Tests/Shared/TestPropagator.cs
… test/OpenTelemetry.Tests/Shared/TestPropagator{TCarrier}.cs
… test/OpenTelemetry.Tests/Trace/Propagation/TraceContextPropagatorTest.cs