## Changes

`ActivityContext` & `PropagationContext` are large structs. Passing them by value means they are copied. The [.NET guidance is to pass large struct by in/ref](https://docs.microsoft.com/en-us/dotnet/csharp/write-safe-efficient-code#use-in-parameters-for-large-structs). This is especially noticeable with the composite propagator which makes a lot of copies as it recurses. My goal with this PR was to pass by ref (`in` or `ref`) in a way that would be backwards compatible with existing propagators/instrumentation.

Before:
|                                      Method |  mode |      Mean |    Error |   StdDev |  Gen 0 | Allocated |
|-------------------------------------------- |------ |----------:|---------:|---------:|-------:|----------:|
|      TraceContextPropagator_InjectBenchmark |     ? |  42.52 ns | 0.659 ns | 0.809 ns | 0.0210 |     264 B |
|                B3Propagator_InjectBenchmark |     ? |  14.27 ns | 0.121 ns | 0.101 ns |      - |         - |
|  CompositeTextMapPropagator_InjectBenchmark |     ? |  76.72 ns | 0.662 ns | 0.553 ns | 0.0210 |     264 B |
|               B3Propagator_ExtractBenchmark |    B3 | 220.19 ns | 2.877 ns | 5.812 ns | 0.0172 |     216 B |
| CompositeTextMapPropagator_ExtractBenchmark |    B3 | 346.86 ns | 1.728 ns | 1.532 ns | 0.0210 |     264 B |
|     TraceContextPropagator_ExtractBenchmark | Empty |  34.67 ns | 0.703 ns | 0.623 ns | 0.0019 |      24 B |
|               B3Propagator_ExtractBenchmark | Empty |  39.79 ns | 0.129 ns | 0.107 ns | 0.0019 |      24 B |
| CompositeTextMapPropagator_ExtractBenchmark | Empty | 123.64 ns | 1.152 ns | 1.131 ns | 0.0057 |      72 B |
|     TraceContextPropagator_ExtractBenchmark |   W3C | 177.61 ns | 2.163 ns | 1.806 ns | 0.0153 |     192 B |
| CompositeTextMapPropagator_ExtractBenchmark |   W3C | 253.52 ns | 1.887 ns | 1.854 ns | 0.0172 |     216 B |

After:
|                                      Method |  mode |      Mean |    Error |   StdDev |  Gen 0 | Allocated |
|-------------------------------------------- |------ |----------:|---------:|---------:|-------:|----------:|
|      TraceContextPropagator_InjectBenchmark |     ? |  39.90 ns | 0.314 ns | 0.278 ns | 0.0210 |     264 B |
|                B3Propagator_InjectBenchmark |     ? |  12.46 ns | 0.189 ns | 0.177 ns |      - |         - |
|  CompositeTextMapPropagator_InjectBenchmark |     ? |  72.59 ns | 0.424 ns | 0.396 ns | 0.0210 |     264 B |
|               B3Propagator_ExtractBenchmark |    B3 | 206.05 ns | 0.503 ns | 0.470 ns | 0.0172 |     216 B |
| CompositeTextMapPropagator_ExtractBenchmark |    B3 | 317.80 ns | 0.824 ns | 0.688 ns | 0.0210 |     264 B |
|     TraceContextPropagator_ExtractBenchmark | Empty |  23.20 ns | 0.067 ns | 0.062 ns | 0.0019 |      24 B |
|               B3Propagator_ExtractBenchmark | Empty |  33.55 ns | 0.115 ns | 0.096 ns | 0.0019 |      24 B |
| CompositeTextMapPropagator_ExtractBenchmark | Empty | 100.98 ns | 0.253 ns | 0.225 ns | 0.0057 |      72 B |
|     TraceContextPropagator_ExtractBenchmark |   W3C | 164.99 ns | 0.414 ns | 0.387 ns | 0.0153 |     192 B |
| CompositeTextMapPropagator_ExtractBenchmark |   W3C | 232.80 ns | 0.501 ns | 0.391 ns | 0.0172 |     216 B |

TODOs:

* [ ] Appropriate `CHANGELOG.md` updated for non-trivial changes
* [ ] Changes in public API reviewed