@alanwest 2022-05-25
The new APIs do not seem as intuitive to use - though this may be pretty subjective. Also the perf benefit doesn't strike me as amazing, but happy to hear other opinions on this.

If we do decide to go forward, I'd be inclined to sit on this at least until after the 1.3.0 release next week. Might give a bit more of an opportunity for folks to give feedback if released with the 1.4.0 beta.