@CodeBlanch 2022-05-25
@reyang @alanwest 

I don't think your instincts are wrong here. I was kind of disappointed the perf gain wasn't greater! Almost threw away the work 🤣 Ultimately decided it was worth opening the PR if only to highlight the effort needed to really get great perf from structs. I'll add a little more color here, but I'm fine with just letting this die on the vine, doing it in the future, or whatever.

_Are there complaints/issues about the current performance, or there is some blocker?_

No

_Would anyone complain if we made it faster?_

Also no 🤣 

_The new APIs do not seem as intuitive to use - though this may be pretty subjective._

Ya. Is what it is. Structs are hard to do correctly 🧙‍♂️ Runtime is [doing](https://github.com/dotnet/runtime/blob/1adf59d5c1e3b41d94695dd9ad11c71fe8823a28/src/libraries/System.Diagnostics.DiagnosticSource/src/System/Diagnostics/ActivityListener.cs#L11) a lot of [similar](https://github.com/dotnet/runtime/blob/1adf59d5c1e3b41d94695dd9ad11c71fe8823a28/src/libraries/System.Diagnostics.DiagnosticSource/src/System/Diagnostics/ActivitySource.cs#L417) things for similar reasons. The one thing with the propagation API, not many people are going to use it. There will probably be a really small set of propagators out there ever. Instrumentation needs to do it correctly. There will probably be more of that, but still not expansive. Basically it isn't something most people will need to worry about it so it might be OK to have a more esoteric API for the perf.

One thing I kind of mentioned on the sig. The benchmarks on the PR are really just covering Inject/Extract. They don't tell the full story! This type of code...

https://github.com/open-telemetry/opentelemetry-dotnet/blob/8aa1778b439e61592fdc5ae238474bb5f2294063/src/OpenTelemetry.Instrumentation.AspNetCore/Implementation/HttpInListener.cs#L89-L114

...uses the propagation API but then also does more with ActivityContext, paying additional copy penalties.

If we expand the benchmarks to cover that additional stuff, more perf gains. I'm using the default W3C + Baggage set up here, but the more propagators the bigger the gain.

Before

<details>

<summary>Code</summary>

```csharp
    public class AspNetCoreExtractionBenchmark
    {
        private const string Traceparent = "00-0af7651916cd43dd8448eb211c80319c-00f067aa0ba902b7-00";
        private static readonly Func<HttpRequest, string, IEnumerable<string>> CarrierGetter = (request, name) => request.Headers[name];
        private static readonly ActivityContext RandomActivityContext = new(ActivityTraceId.CreateRandom(), ActivitySpanId.CreateRandom(), ActivityTraceFlags.Recorded);
        private static readonly ActivityContext MatchActivityContext;
        private static readonly TextMapPropagator Propagator = new CompositeTextMapPropagator(new TextMapPropagator[]
        {
            new TraceContextPropagator(),
            new BaggagePropagator(),
        });

        private readonly HttpRequest w3cExtractRequest;

        static AspNetCoreExtractionBenchmark()
        {
            var parsedContext = ActivityContext.Parse(Traceparent, traceState: null);
            MatchActivityContext = new ActivityContext(
                parsedContext.TraceId,
                parsedContext.SpanId,
                parsedContext.TraceFlags,
                parsedContext.TraceState,
                isRemote: true);
        }

        public AspNetCoreExtractionBenchmark()
        {
            HttpContext w3cContext = new DefaultHttpContext();
            this.w3cExtractRequest = w3cContext.Request;
            this.w3cExtractRequest.Headers.Add(TraceContextPropagator.TraceParent, Traceparent);
        }

        [Benchmark]
        [Arguments(true)]
        [Arguments(false)]
        public Activity AspNetCoreExtractBenchmark(bool match)
        {
            ActivityContext compareContext = match ? MatchActivityContext : RandomActivityContext;
            Activity createdActivity = null;

            var textMapPropagator = Propagator;
            if (textMapPropagator is not TraceContextPropagator)
            {
                var ctx = textMapPropagator.Extract(default, this.w3cExtractRequest, CarrierGetter);

                if (ctx.ActivityContext.IsValid()
                    && ctx.ActivityContext != compareContext)
                {
                    createdActivity = new Activity("InnerActivity");
                    createdActivity.SetParentId(ctx.ActivityContext.TraceId, ctx.ActivityContext.SpanId, ctx.ActivityContext.TraceFlags);
                    createdActivity.TraceStateString = ctx.ActivityContext.TraceState;
                }

                Baggage.Current = ctx.Baggage;
            }

            return createdActivity;
        }
    }
```

</details>

|                     Method | match |     Mean |   Error |  StdDev |  Gen 0 | Allocated |
|--------------------------- |------ |---------:|--------:|--------:|-------:|----------:|
| AspNetCoreExtractBenchmark | False | 286.8 ns | 1.88 ns | 1.66 ns | 0.0324 |     408 B |
| AspNetCoreExtractBenchmark |  True | 271.8 ns | 2.99 ns | 2.80 ns | 0.0172 |     216 B |

After

<details>

<summary>Code</summary>

```csharp
    public class AspNetCoreExtractionBenchmark
    {
        private const string Traceparent = "00-0af7651916cd43dd8448eb211c80319c-00f067aa0ba902b7-00";
        private static readonly Func<HttpRequest, string, IEnumerable<string>> CarrierGetter = (request, name) => request.Headers[name];
        private static readonly ActivityContext RandomActivityContext = new(ActivityTraceId.CreateRandom(), ActivitySpanId.CreateRandom(), ActivityTraceFlags.Recorded);
        private static readonly ActivityContext MatchActivityContext = ActivityContext.Parse(Traceparent, traceState: null);
        private static readonly TextMapPropagator Propagator = new CompositeTextMapPropagator(new TextMapPropagator[]
        {
            new TraceContextPropagator(),
            new BaggagePropagator(),
        });

        private readonly HttpRequest w3cExtractRequest;

        public AspNetCoreExtractionBenchmark()
        {
            HttpContext w3cContext = new DefaultHttpContext();
            this.w3cExtractRequest = w3cContext.Request;
            this.w3cExtractRequest.Headers.Add(TraceContextPropagator.TraceParent, Traceparent);
        }

        [Benchmark]
        [Arguments(true)]
        [Arguments(false)]
        public Activity AspNetCoreExtractBenchmark(bool match)
        {
            ActivityContext compareContext = match ? MatchActivityContext : RandomActivityContext;
            Activity createdActivity = null;

            var textMapPropagator = Propagator;
            if (textMapPropagator is not TraceContextPropagator)
            {
                PropagationContext ctx = default;
                textMapPropagator.Extract(ref ctx, this.w3cExtractRequest, CarrierGetter);

                ref readonly ActivityContext activityContext = ref PropagationContext.GetActivityContextRef(in ctx);

                if (ActivityContextExtensions.IsValid(in activityContext)
                    && (!activityContext.TraceId.Equals(compareContext.TraceId)
                    || !activityContext.SpanId.Equals(compareContext.SpanId)
                    || activityContext.TraceFlags != compareContext.TraceFlags
                    || activityContext.TraceState != compareContext.TraceState))
                {
                    createdActivity = new Activity("InnerActivity");
                    createdActivity.SetParentId(activityContext.TraceId, activityContext.SpanId, activityContext.TraceFlags);
                    createdActivity.TraceStateString = activityContext.TraceState;
                }

                Baggage.Current = ctx.Baggage;
            }

            return createdActivity;
        }
    }
```

</details>

|                     Method | match |     Mean |   Error |  StdDev |  Gen 0 | Allocated |
|--------------------------- |------ |---------:|--------:|--------:|-------:|----------:|
| AspNetCoreExtractBenchmark | False | 264.3 ns | 1.24 ns | 1.10 ns | 0.0324 |     408 B |
| AspNetCoreExtractBenchmark |  True | 250.5 ns | 1.47 ns | 1.31 ns | 0.0172 |     216 B |

In that case we pick up some nanos on the root. We'll likely pick up some nanos on outgoing spans as well.

Worth it? 🤷‍♀️